Split renderers into separate module
