+++ chrome-ext/src/manifest.json
@@ -31,7 +31,8 @@
       "js": [
         "js/tingle.js",
         "js/mustache.js",
-        "js/hots-dialog.js"
+        "js/hots-dialog.js",
+        "js/hots-dialog-renderers.js"
       ]
     },
     {
